+++ c-ext/python-zstandard.h
@@ -86,10 +86,13 @@ typedef struct {
 	void* dictData;
 	/* Size of dictionary data. */
 	size_t dictSize;
+	ZSTD_dictMode_e dictMode;
 	/* k parameter for cover dictionaries. Only populated by train_cover_dict(). */
 	unsigned k;
 	/* d parameter for cover dictionaries. Only populated by train_cover_dict(). */
 	unsigned d;
+	/* Digested dictionary, suitable for reuse. */
+	ZSTD_CDict* cdict;
 } ZstdCompressionDict;
 
 extern PyTypeObject ZstdCompressionDictType;