+++ tests/test_module_attributes.py
@@ -51,6 +51,9 @@ class TestModuleAttributes(unittest.TestCase):
             'STRATEGY_BTLAZY2',
             'STRATEGY_BTOPT',
             'STRATEGY_BTULTRA',
+            'DICT_MODE_AUTO',
+            'DICT_MODE_RAWCONTENT',
+            'DICT_MODE_FULLDICT',
         )
 
         for a in attrs: